openhab2/coap-bridge/requirements.txt
aiocoap>=0.4.0
paho-mqtt>=2.0.0
